Tested SerialTransfer library
I had a chance to test the SerialTransfer library and can confirm that it works on the Pico. Updated the "test" column to reflect this.

I noticed that after my last change to this list was merged, the page failed to build. Is there something else I need to do in order to update this list?